+++ source/collect/CollectService.ts
@@ -11,55 +11,18 @@ import { TestTree } from "./TestTree.js";
 import { TestTreeNode } from "./TestTreeNode.js";
 import { TestTreeNodeBrand } from "./TestTreeNodeBrand.enum.js";
 import { TestTreeNodeFlags } from "./TestTreeNodeFlags.enum.js";
-import type { SuppressedError, SuppressedErrors } from "./types.js";
 import { WhenNode } from "./WhenNode.js";
 
 export class CollectService {
   #abilityLayer: AbilityLayer;
   #compiler: typeof ts;
-  #expectErrorRegex = /^( *)(\/\/ *@ts-expect-error)(!?)(:? *)(.*)?$/gim;
   #identifierLookup: IdentifierLookup;
-  #resolvedConfig: ResolvedConfig;
 
   constructor(compiler: typeof ts, projectService: ProjectService, resolvedConfig: ResolvedConfig) {
     this.#compiler = compiler;
-    this.#resolvedConfig = resolvedConfig;
 
     this.#abilityLayer = new AbilityLayer(compiler, projectService, resolvedConfig);
     this.#identifierLookup = new IdentifierLookup(compiler);
-  }
-
-  #collectSuppressedErrors(sourceFile: ts.SourceFile) {
-    const ranges: SuppressedErrors = Object.assign([], { sourceFile });
-
-    for (const match of sourceFile.text.matchAll(this.#expectErrorRegex)) {
-      const offsetText = match?.[1];
-      const directiveText = match?.[2];
-      const ignoreText = match?.[3];
-      const argumentSeparatorText = match?.[4];
-      const argumentText = match?.[5]?.split(/--+/)[0]?.trimEnd();
-
-      if (typeof offsetText !== "string" || !directiveText || ignoreText === "!") {
-        continue;
-      }
-
-      const start = match.index + offsetText.length;
-
-      const range: SuppressedError = {
-        directive: { start, end: start + directiveText.length, text: directiveText },
-        diagnostics: [],
-      };
-
-      if (typeof argumentSeparatorText === "string" && typeof argumentText === "string") {
-        const start = range.directive.end + argumentSeparatorText.length;
-
-        range.argument = { start, end: start + argumentText.length, text: argumentText };
-      }
-
-      ranges.push(range);
-    }
-
-    return ranges;
   }
 
   #collectTestTreeNodes(node: ts.Node, parent: TestTree | TestTreeNode, testTree: TestTree) {
@@ -190,14 +153,8 @@ export class CollectService {
 
     EventEmitter.dispatch(["collect:start", { tree: testTree }]);
 
-    this.#abilityLayer.open(sourceFile);
+    this.#abilityLayer.open(testTree);
     this.#identifierLookup.open();
-
-    if (this.#resolvedConfig.checkSuppressedErrors) {
-      testTree.suppressedErrors = this.#collectSuppressedErrors(sourceFile);
-
-      this.#abilityLayer.handleSuppressedErrors(testTree);
-    }
 
     this.#collectTestTreeNodes(sourceFile, testTree, testTree);
